requirements.txt
appdirs==1.4.3
beowulf==1.0.2
certifi==2019.3.9
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.1
django-dotenv==1.4.2
django-rest-swagger==2.2.0
djangorestframework==3.9.3
ecdsa==0.13.2
funcy==1.12
future==0.17.1
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
langdetect==1.0.7
MarkupSafe==1.1.1
openapi-codec==1.3.2
prettytable==0.7.2
psycopg2-binary==2.8.2
pycrypto==2.6.1
pylibscrypt==1.8.0
pytz==2019.1
requests==2.22.0
scrypt==0.8.13
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
toolz==0.9.0
ujson==1.35
uritemplate==3.0.0
urllib3==1.25.2
voluptuous==0.11.5
w3lib==1.20.0
